backend/requirements.txt
aiohttp==3.8.4
aiohttp-middlewares==2.3.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
dataclasses-json==0.5.9
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-speech==2.21.0
googleapis-common-protos==1.59.1
greenlet==1.1.2
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.233
langsmith==0.0.5
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydantic==1.10.11
python-engineio==4.3.2
python-socketio==5.6.0
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.18
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.0.0
yarl==1.9.2
zipp==3.16.1
